// ghd-expo/components/dashboard/types.ts
export type TripSummary = {
    distance: string;
    remaining: string;
    cumulativeEnergy: string;
    maxSpeed: string;
    avgSpeed: string;
    avgPower: string;
    whPerUnit: string;
    startTime: number;
    maxVoltageSag: string;
    gpsMaxSpeed: string;
    gpsAvgSpeed: string;
    gpsSampleCount: number;
};